+++ src/main/java/com/telepathicgrunt/the_bumblezone/blocks/StickyHoneyResidue.java
@@ -70,19 +70,19 @@ public class StickyHoneyResidue extends VineBlock {
             voxelshape = VoxelShapes.or(voxelshape, VineBlockAccessor.thebumblezone_getUP_SHAPE());
         }
 
-        if (blockstate.getValue(SOUTH)) {
+        if (blockstate.getValue(NORTH)) {
             voxelshape = VoxelShapes.or(voxelshape, VineBlockAccessor.thebumblezone_getNORTH_SHAPE());
         }
 
-        if (blockstate.getValue(WEST)) {
+        if (blockstate.getValue(EAST)) {
             voxelshape = VoxelShapes.or(voxelshape, VineBlockAccessor.thebumblezone_getEAST_SHAPE());
         }
 
-        if (blockstate.getValue(NORTH)) {
+        if (blockstate.getValue(SOUTH)) {
             voxelshape = VoxelShapes.or(voxelshape, VineBlockAccessor.thebumblezone_getSOUTH_SHAPE());
         }
 
-        if (blockstate.getValue(EAST)) {
+        if (blockstate.getValue(WEST)) {
             voxelshape = VoxelShapes.or(voxelshape, VineBlockAccessor.thebumblezone_getWEST_SHAPE());
         }